Bug 522018 - Rewrite prefs endpoint, and fix all failures from test-prefs.js
