+++ modules/orionode/index.js
@@ -119,7 +119,7 @@ function startServer(options) {
 		app.use('/workspace*', options.authenticate, checkAuthenticated, checkAccessRights, require('./lib/workspace')(options));
 		app.use('/gitapi', options.authenticate, checkAuthenticated, require('./lib/git')(options));
 		app.use('/cfapi', options.authenticate, checkAuthenticated, require('./lib/cf')(options));
-		app.use('/prefs', options.authenticate, checkAuthenticated, require('./lib/controllers/prefs').router(options));
+		app.use('/prefs', options.authenticate, checkAuthenticated, require('./lib/prefs').router(options));
 		app.use('/xfer', options.authenticate, checkAuthenticated, require('./lib/xfer').router(options));
 		if(options.configParams["orion.collab.enabled"]){
 			app.use('/sharedWorkspace', options.authenticate, checkAuthenticated, require('./lib/sharedWorkspace').router(options));
