+++ modules/orionode/lib/update.js
@@ -14,7 +14,7 @@ var express = require('express'),
 	electron = require('electron'),
 	autoUpdater = require('./autoUpdater.js'),
 	bodyParser = require('body-parser'),
-	prefs = require('./controllers/prefs'),
+	prefs = require('./prefs'),
 	os = require('os'),
 	platform = os.platform(),
 	arch = os.arch(),
